Move get_search_suggestions from browsing to search.
